--- API/callAnki.py
@@ -1,24 +1,0 @@
-import urllib.request
-import json
-
-
-
-class callAnki:
-    def __init__(self, deck_name):
-        self.deck_name = deck_name
-
-    def request(action, **params):
-        return {'action': action, 'params': params, 'version': 6}
-
-    def invoke(action, **params):
-        requestJson = json.dumps(urllib.request.request(action, **params)).encode('utf-8')
-        response = json.load(urllib.request.urlopen(urllib.request.Request('http://127.0.0.1:8765', requestJson)))
-        if len(response) != 2:
-            raise Exception('response has an unexpected number of fields')
-        if 'error' not in response:
-            raise Exception('response is missing required error field')
-        if 'result' not in response:
-            raise Exception('response is missing required result field')
-        if response['error'] is not None:
-            raise Exception(response['error'])
-        return response['result']
